+++ wls-gui-wahllokalsystem/src/components/ergebnismeldung/OBW/stapelA/BaseRowOBWStapelA.vue
