+++ wls-gui-wahllokalsystem/src/components/ergebnismeldung/OBW/stapelB/TheOBWStapelBCard.vue
