+++ wls-gui-wahllokalsystem/src/components/ergebnismeldung/OBW/stapelC/BaseRowStapelC.vue
