+++ wls-gui-wahllokalsystem/src/components/ergebnismeldung/OBW/stapelC/TheOBWStapelCDeletionDeniedDialog.vue
