+++ wls-gui-wahllokalsystem/src/components/ergebnismeldung/OBW/stapelC/TheOBWStapelCSummaryCard.vue
