+++ wls-gui-wahllokalsystem/stories/components/ergebnismeldung/OBW/stapelC/TheOBWStapelCSummaryCard.stories.ts
@@ -3,7 +3,7 @@ import type { Meta, StoryObj } from "@storybook/vue3";
 import { useErgebnisseTestDataFactory } from "@tests/utils/ergebnismeldung/ergebnisseTestDataFactory.ts";
 import { useWahlvorschlaegeTestDataFactory } from "@tests/utils/wahlvorschlaege/WahlvorschlaegeTestDataFactory.ts";
 
-import TheOBWStapelCSummaryCard from "@/components/ergebnisermittlung/OBW/stapelC/TheOBWStapelCSummaryCard.vue";
+import TheOBWStapelCSummaryCard from "@/components/ergebnismeldung/OBW/stapelC/TheOBWStapelCSummaryCard.vue";
 import pinia from "@/plugins/pinia.ts";
 import { useErgebnismeldungStore } from "@/stores/ergebnismeldungStore.ts";
 import { useWahlvorschlaegeStore } from "@/stores/wahlvorschlaegeStore.ts";
